+++ CHANGES.md
@@ -3,6 +3,7 @@
 * Added support for the PDE editor.
 * Added support for the ColdFusion editor (CFEclipse).
 * Added support for the TCL editor.
+* Added support for the TeXlipse editor.
 * Fixed some color mappings in C/C++ editor.
 * Fixed some color values in Oblivion theme.